+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Breaking:** Page number placeholder (or index key) has been changed
+  from `i` to `p`. For example, you use `output_{p}.json` now.
 - **Breaking:** Make `--data` in `generate typst --data` a positional argument
   with default value instead of an option.